refactor: MyGatheringCard를 MyGahteringSection 하위 폴더로 이동하여 Colocation 적용
