+++ src/app/main/components/MyGatheringSection/MyGatheringCard/index.tsx
